Successully creted the UI for frontend/Added data successfully

+++ backend/controllers/product.controller.js
@@ -46,13 +46,18 @@ export const updateProduct = async (req, res) => {
   }
 };
 
-export const deleteProduct =  async (req, res) => {
-    const { id } = req.params;
-    try {
-      await Product.findByIdAndDelete(id);
-      res.status(200).json({ success: true, message: "Product deleted" });
-    } catch (error) {
-      console.log("eror in deleting product:", error.message);
-      res.status(404).json({ success: false, message: "Product not found" });
-    }
-  };
+export const deleteProduct = async (req, res) => {
+  const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+  try {
+    await Product.findByIdAndDelete(id);
+    res.status(200).json({ success: true, message: "Product deleted" });
+  } catch (error) {
+    console.log("eror in deleting product:", error.message);
+    res.status(500).json({ success: false, message: "Server Error" });
+  }
+};
